requirements.txt
Flask==2.2.2
google_api_python_client==2.71.0
protobuf==4.21.12
python-dotenv==0.21.0
requests==2.22.0
